+++ app/controllers/accounts_controller.rb
@@ -1,107 +1,101 @@
 class AccountsController < ApplicationController
-  before_filter :basic, :only => [:register]
+  before_filter :basic, :only => [:manage]
 
-  def register
+  def manage
     @account = Account.new
-    @all_accounts = Account.all
+    @accounts = Account.order(:date => :desc).page(params[:page])
   end
 
   def create
-    params.permit!
-    check_absent_params_for_create
-
     begin
-      accounts = params[:accounts].slice(*account_attributes)
-      accounts[:date] = 'invalid_date' unless accounts[:date] =~ /\d{4}-\d{2}-\d{2}/
-      @account = Account.create!(accounts)
-      if params[:accounts][:from] == 'browser'
-        render
+      attributes = params.require(:accounts).permit(*account_params)
+      absent_keys = account_params - attributes.symbolize_keys.keys
+      raise BadRequest.new(absent_keys, 'absent') unless absent_keys.empty?
+
+      @account = Account.new(attributes)
+      if @account.save
+        request.format = :js if request.xhr?
+        respond_to do |format|
+          format.json {render :status => :created, :template => 'accounts/account'}
+          format.js {@accounts = Account.order(:date => :desc).page(params[:page])}
+        end
       else
-        render :status => :created, :json => @account
+        raise BadRequest.new(@account.errors.messages.keys, 'invalid')
       end
-    rescue ActiveRecord::RecordInvalid => e
-      raise BadRequest.new(e.record.errors.messages.keys, 'invalid')
+    rescue ActionController::ParameterMissing
+      raise BadRequest.new(:accounts, 'absent')
     end
   end
 
-  def read
-    params.permit!
+  def show
+    @account = Account.find_by(params.permit(:id))
+    if @account
+      respond_to do |format|
+        format.json {render :status => :ok, :template => 'accounts/account'}
+      end
+    else
+      raise NotFound.new
+    end
+  end
 
-    begin
-      render :status => :ok, :json => Account.show(params.slice(*account_attributes))
-    rescue ActiveRecord::RecordInvalid => e
-      raise BadRequest.new(e.record.errors.messages.keys, 'invalid')
+  def index
+    query = Query.new(params.permit(*index_params))
+    if query.valid?
+      @accounts = index_params.inject(Account.all) do |accounts, key|
+        value = query.send(key)
+        value ? accounts.send(key, value) : accounts
+      end
+      respond_to do |format|
+        format.json {render :status => :ok, :template => 'accounts/accounts'}
+      end
+    else
+      raise BadRequest.new(query.errors.messages.keys, 'invalid')
     end
   end
 
   def update
-    params.permit!
-    check_absent_params_for_update
-
-    begin
-      render :status => :ok, :json => Account.update(params.slice(*permitted_params_update))
-    rescue ActiveRecord::RecordInvalid => e
-      raise BadRequest.new(e.record.errors.messages.keys, 'invalid')
+    @account = Account.find_by(params.permit(:id))
+    if @account
+      if @account.update(params.permit(*account_params))
+        respond_to do |format|
+          format.json {render :status => :ok, :template => 'accounts/account'}
+        end
+      else
+        raise BadRequest.new(@account.errors.messages.keys, 'invalid')
+      end
+    else
+      raise NotFound.new
     end
   end
 
-  def delete
-    params.permit!
-
-    begin
-      Account.destroy params.slice(*account_attributes)
-      render :status => :no_content, :nothing => true
-    rescue ActiveRecord::RecordInvalid => e
-      raise BadRequest.new(e.record.errors.messages.keys, 'invalid')
+  def destroy
+    @account = Account.find_by(params.permit(:id)).try(:destroy)
+    if @account
+      head :no_content
+    else
+      raise NotFound.new
     end
   end
 
   def settle
-    params.permit!
-    check_absent_params_for_settle
-
-    begin
-      render :status => :ok, :json => Account.settle(params[:interval])
-    rescue Exception
-      raise BadRequest.new(:interval, 'invalid')
+    query = Settlement.new(params.permit(:interval))
+    if query.valid?
+      @settlement = Account.settle(query.interval)
+      respond_to do |format|
+        format.json {render :status => :ok}
+      end
+    else
+      raise BadRequest.new(:interval, query.errors.messages[:interval].first)
     end
   end
 
   private
 
-  def account_attributes
+  def account_params
     %i[ account_type date content category price ]
   end
 
-  def permitted_params_update
-    %i[ condition with ]
-  end
-
-  def permitted_values_settle
-    %i[ yearly monthly daily ]
-  end
-
-  def check_absent_params_for_create
-    request_params = request.request_parameters
-    raise BadRequest.new(:accounts, 'absent') unless request_params.has_key?(:accounts)
-    errors = [].tap do |array|
-      account_attributes.each do |param_key|
-        array << param_key unless request_params[:accounts].has_key?(param_key)
-      end
-    end
-    raise BadRequest.new(errors, 'absent') unless errors.empty?
-  end
-
-  def check_absent_params_for_update
-    request_params = request.request_parameters
-    raise BadRequest.new(:with, 'absent') unless request_params.has_key?(:with)
-    raise BadRequest.new(:with, 'absent') if request_params[:with].empty?
-  end
-
-  def check_absent_params_for_settle
-    raise BadRequest.new(:interval, 'absent') unless request.query_parameters.has_key?(:interval)
-    unless request.query_parameters[:interval].match(/#{permitted_values_settle.join('|')}/)
-      raise BadRequest.new(:interval, 'invalid')
-    end
+  def index_params
+    %i[ account_type date_before date_after content_equal content_include category price_upper price_lower ]
   end
 end