+++ app/models/settlement.rb
@@ -1,0 +1,8 @@
+class Settlement
+  include ActiveModel::Model
+
+  attr_accessor :interval
+
+  validates :interval, :presence => {:message => 'absent'}
+  validates :interval, :inclusion => {:in => %w[ daily monthly yearly ], :message => 'invalid'}, :allow_nil => true
+end