+++ config/routes.rb
@@ -1,10 +1,5 @@
 Rails.application.routes.draw do
-  post '/accounts' => 'accounts#create'
-  get '/accounts' => 'accounts#read'
-  put '/accounts' => 'accounts#update'
-  delete '/accounts' => 'accounts#delete'
-
-  get '/settlement' => 'accounts#settle'
-
-  get '/' => 'accounts#register'
+  resources :accounts, :except => [:new, :edit], :defaults => {:format => 'json'}
+  get '/settlement' => 'accounts#settle', :defaults => {:format => 'json'}
+  get '/' => 'accounts#manage'
 end