+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 algieba
 =======
 
+- 2.1.3:
+  - [FIX] default format
+
+- 2.1.2:
+  - [FIX] XHR error
+  - [CHANGE] input width for content
+  - [REFACTOR] error handling
+
+- 2.1.1:
+  - [FIX] routing error for assets
+
+- 2.1.0:
+  - [ADD] pagination
+
+- 2.0.0:
+  - [ADD] indexing api
+
 - 1.1.1:
   - [ADD] exception class
   - [FIX] request spec